% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_placebos.R
\name{create_placebos}
\alias{create_placebos}
\title{Multiple placebocheks unified in just one list or coefplot}
\usage{
create_placebos(data, cutoff, formula, operations, coefplot = F, geometry = F)
}
\arguments{
\item{data}{sf data.frame that contains all units of observation}

\item{cutoff}{initial RD cutoff as an sj object}

\item{formula}{provide the formula you want to use for OLS, omit the treatetment dummy (if you want a univariate regression just on "treated", then provide y ~ 1 as formula)}

\item{operations}{container that has all the information in it on how to change the border for each placeboregression}

\item{coefplot}{provide coefplot instead of a data.frame}

\item{geometry}{set to \code{TRUE} if you want to plot all the lines of the used placebo borders}
}
\value{
either a coefplot or data.frame containing results of placeboregressions
}
\description{
Unifies \code{\link{shift_border}}, \code{\link{cutoff2polygon}}, \code{\link{assign_treated}} in one function to carry out a myriad of placebo checks at once.
The output is either a data.frame (with or without geometry of the respective placeboline) or a coefplot.
Requires operations data.frame that contains all desired operations (columns shift.x, shift.y, scale, angle, orientation.1, orientation.2, endpoint.1, endpoint.2),
if you don't need a certain operation just use default values (e.g. 0 for angle and 1 for scale), but the column has to be there.
}
\examples{
\dontrun{create_placebos(data = points_samp.sf, cutoff = cut_off.sf,
formula = education ~ 1, operations = operations)}
}
