% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cut_off}
\alias{cut_off}
\title{Dataset with boundaries and polygons for the SpatialRDD vignette.}
\format{
A spatial data.frame of class sf
}
\source{
Lehner, Alexander (2023) Culture, Institutions, and the Roots of Gender Inequality: 450 Years of Portuguese Colonialism in India
}
\usage{
data(cut_off)
}
\description{
sf multilinestring representing a spatial RD cut-off
}
\keyword{datasets}
