% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cut_off.sf}
\alias{cut_off.sf}
\title{Dataset with boundaries and polygons for the SpatialRDD vignette.}
\format{
A spatial data.frame of class sf
}
\source{
Lehner, A. (2019) 450 Years of Portuguese Colonialism in India: Missionaries, Education, and the Roots of Gender Inequality
}
\usage{
data(cut_off.sf)
}
\description{
sf multilinestring representing a spatial RD cut-off
}
\keyword{datasets}
