% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff2polygon.R
\name{cutoff2polygon}
\alias{cutoff2polygon}
\title{Create (treated) polygon from line}
\usage{
cutoff2polygon(
  data,
  cutoff,
  orientation = NA,
  endpoints = c(0, 0),
  corners = NA,
  difference = F,
  upside = T
)
}
\arguments{
\item{data}{study dataset to determine the bounding box (so that all observations are covered by the new polygons) in sf format}

\item{cutoff}{sf object of the (placebo) cut-off}

\item{orientation}{how is the discontinuity oriented (e.g. "west-west" means both ends of the line go to the western end of the bounding box)}

\item{endpoints}{at what position on the edge should each polygon end (between 0 and 1, where 0.5 e.g. means right in the middle of the respective edge)}

\item{corners}{clockwise 1-4 (numbers refer to quadrant corners i.e. 1 is top right). 2 corners max. If you need 3 go the other way round with 1 and take the spatial difference with the bounding box afterwards.}

\item{upside}{deprecated, this is replaced by doing an intersect with the bounding box to get the difference}

\item{crs}{epsg number if project crs}
}
\value{
a polygon as an sf object
}
\description{
Creates an approximation of a "treated/untreated polygon" to assign the status again to each observation after the border has been shifted
}
