% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff2polygons.R
\name{cutoff2polygons}
\alias{cutoff2polygons}
\title{Create treated polygon from (placebo) line}
\usage{
cutoff2polygons(data = bbox, cutoff = lines.sf, orientation = NA,
  endpoints = c(0, 0), upside = T, crs = crs)
}
\arguments{
\item{data}{study dataset to determine the bounding box (so that all observations are covered by the new polygons) in sf format}

\item{cutoff}{sf object of the (placebo) cut-off}

\item{orientation}{how is the discontinuity oriented (e.g. "west-west" means both ends of the line go to the western end of the bounding box)}

\item{endpoints}{at what position on the edge should each polygon end (between 0 and 1, where 0.5 e.g. means right in the middle of the respective edge)}

\item{upside}{upper- or lowerside? run one with T and once with F to cover the full studyarea (e.g. treated and untreated)}

\item{crs}{epsg number if project crs}
}
\description{
Creates a very crude approximation of a "treated/untreated polygon" that makes it easier to assign the status again to each observation after the border has been shifted
}
