% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise_border.R
\name{discretise_border}
\alias{discretise_border}
\title{This function takes in a border in the form of a polyline
and converts it into point data
these points are later used to calculate a heterogenous treatment effect alongside the border (i.e. the cut-off)}
\usage{
discretise_border(cutoff = cut_off.sf, n = 10, random = F)
}
\description{
supported formats for the line:
}
