% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_border.R
\name{placebo_border}
\alias{placebo_border}
\title{Title}
\usage{
placebo_border(border = cut_off.sf, operation = c("shift", "scale",
  "rotate"), shift = c(0, 0), scale = 1, angle = 0)
}
\arguments{
\item{angle}{}
}
\description{
Title
}
