% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspatialrd.R
\name{plotspatialrd}
\alias{plotspatialrd}
\title{Plot SpatialRD output}
\usage{
plotspatialrd(SpatialRDoutput = results, map = T)
}
\arguments{
\item{SpatialRDoutput}{spatial obkect that is produce by an estimation with \code{\link{SpatialRD}}}

\item{map}{T/F depending on whether mapplot is desired}
}
\value{
plots produced with ggplot 2
}
\description{
Produces plot of GRDDseries and optionally of a map that visualises every point estimate in space
}
