% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2line.R
\name{points2line}
\alias{points2line}
\title{Title
this is just a test function, export was removed for now}
\usage{
points2line(borderpoints, crs)
}
\arguments{
\item{borderpoints}{a set of points on a boundary}

\item{crs}{set the coordinate reference system (CRS)}
}
\value{
a line as an sf object
}
\description{
Title
this is just a test function, export was removed for now
}
\examples{
\dontrun{test}
}
