% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygon_full.sf}
\alias{polygon_full.sf}
\title{Dataset with boundaries and polygons for the SpatialRDD vignette.}
\format{A spatial data.frame of class sf}
\source{
Lehner, A. (2019) 450 Years of Portuguese Colonialism in India: Missionaries, Education, and the Roots of Gender Inequality
}
\usage{
data(polygon_full.sf)
}
\description{
sf multipolygon
}
\keyword{datasets}
