% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygon_treated}
\alias{polygon_treated}
\title{Dataset with boundaries and polygons for the SpatialRDD vignette.}
\format{
A spatial data.frame of class sf
}
\source{
Lehner, Alexander (2023) Culture, Institutions, and the Roots of Gender Inequality: 450 Years of Portuguese Colonialism in India
}
\usage{
data(polygon_treated)
}
\description{
sf multipolygon
}
\keyword{datasets}
