% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printspatialrd.R
\name{printspatialrd}
\alias{printspatialrd}
\title{Print SpatialRD output}
\usage{
printspatialrd(SpatialRDoutput = results, label = NA, caption = NA,
  footnote = NA, McCrary = F, RATest = F)
}
\arguments{
\item{footnote}{}
}
\description{
preliminary function using kable and kableextra
}
