% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRD.R
\name{SpatialRD}
\alias{SpatialRD}
\title{This is essentially the main function
It loops over all boundary points and locally estimates a non-parametric RD (local linear regression as usual)
using the rdrobust function from the rdrobust package from Calonico, Cattaneo, Titiunik (2014) Econometrica}
\usage{
SpatialRD(y = "regressand", data = points.sf,
  cutoff.points = borderpoints.sf, treated = "treated", minobs = 50,
  bwfix = F, bwfix_m = NA, sample = F, samplesize = 10,
  cluster = NA)
}
\arguments{
\item{y}{A number.}

\item{x}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
It takes in the discretised cut-off point file (the RDcut-off linestring chopped into parts by the SpatialRDD::discretise_border() function)
and the sf object (which essentially is just a conventional data.frame with a geometry() column) containing all the observations (treated and untreated)
the treated dummy variable has to be assigned before and be part of the sf object as a column
distances etc are all computed within the function using sf
}
\details{
“An object is data with functions. A closure is a function with data.” — John D. Cook
}
