% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treated.R
\name{assign_treated}
\alias{assign_treated}
\title{Title}
\usage{
assign_treated(data.sf = points.sf, polygon.sf = polygon_treated.sf,
  id = "id_column")
}
\arguments{
\item{id}{}
}
\description{
Title
}
