% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspatialrd.R
\name{plotspatialrd}
\alias{plotspatialrd}
\title{Plot SpatialRD output}
\usage{
plotspatialrd(SpatialRDoutput, map = FALSE)
}
\arguments{
\item{SpatialRDoutput}{spatial object that is produced by an estimation with \code{\link{spatialrd}}}

\item{map}{TRUE/FALSE depending on whether mapplot is desired (make sure to set \code{spatial.objcet = TRUE} in the \code{\link{spatialrd}} function)}
}
\value{
plots produced with ggplot2
}
\description{
Produces plot of GRDDseries and optionally of a map that visualises every point estimate in space.
}
\examples{
\donttest{plotspatialrd(results.spatialrd)}
}
