% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2line.R
\name{points2line}
\alias{points2line}
\title{Convert borderpoints to a line}
\usage{
points2line(borderpoints, crs)
}
\arguments{
\item{borderpoints}{a set of points on a boundary}

\item{crs}{set the coordinate reference system (CRS)}
}
\value{
a line as an sf object
}
\description{
Small function that connects dots and makes them one line which can later be used as a cutoff for the RD.
}
\examples{
points_samp.sf <- sf::st_sample(polygon_full, 2) # create points
# make it an sf object bc st_sample only created the geometry list-column (sfc):
points_samp.sf <- sf::st_sf(points_samp.sf)
points2line(points_samp.sf, crs = sf::st_crs(points_samp.sf))

}
