% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygon_full.sf}
\alias{polygon_full.sf}
\title{Dataset with boundaries and polygons for the SpatialRDD vignette.}
\format{A data frame with 3 sf objects:
\describe{
  \item{test}{text}
  \item{test}{text}
  ...
}}
\source{
Lehner2019
}
\usage{
polygon_full.sf
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
