% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printspatialrd.R
\name{printspatialrd}
\alias{printspatialrd}
\title{Print spatialrd output}
\usage{
printspatialrd(SpatialRDoutput = results, format = "latex")
}
\arguments{
\item{SpatialRDoutput}{output file from the \code{\link{spatialrd}} function}

\item{format}{for now only latex}
}
\value{
A formatted table with results from the \code{\link{spatialrd}} function
}
\description{
Preliminary function using kable and kableextra.
You could also just use the package of your choice to print out columns of the output from \code{\link{spatialrd}}.
}
\examples{
printspatialrd(results.spatialrd)
}
