% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRD.R
\name{spatialrd}
\alias{spatialrd}
\title{non-parametric Spatial RD / GRD}
\usage{
spatialrd(
  y,
  data,
  cutoff.points,
  treated,
  minobs = 50,
  bwfix_m = NA,
  sample = FALSE,
  samplesize = NA,
  sparse.exclusion = FALSE,
  store.CIs = FALSE,
  spatial.object = TRUE,
  ...
)
}
\arguments{
\item{y}{The name of the dependent variable in the points frame in the form of a string}

\item{data}{sf data.frame with points that describe the observations}

\item{cutoff.points}{sf object of borderpoints (provided by user or obtained with \code{\link{discretise_border}})}

\item{treated}{column that contains the treated dummy (as string)}

\item{minobs}{the minimum amount of observations in each estimation for the point estimate to be included (default is 50)}

\item{bwfix_m}{fixed bandwidth in meters (in case you want to impose one yourself)}

\item{sample}{draw a random sample of points (default is FALSE)}

\item{samplesize}{if random, how many points}

\item{sparse.exclusion}{in case we want to try to exclude sparse border points before the estimation (should reduce warnings)}

\item{store.CIs}{set TRUE of confidence intervals should be stored}

\item{spatial.object}{return a spatial object (deafult is TRUE, needed if you want to plot the point estimates on a map)?}

\item{...}{in addition you can use all options in \code{\link[rdrobust]{rdrobust}}}
}
\value{
a data.frame or spatial data.frame (sf object) in case spatial.object = TRUE (default)
}
\description{
This function loops over all boundary points and locally estimates a non-parametric RD (local linear regression, as usual)
using the \code{rdrobust} function from the \code{rdrobust} package from Calonico, Cattaneo, Titiunik (2014).
It takes in the discretized cutoff point file (the RDcutoff, a linestring chopped into parts by the \code{\link{discretise_border}} function)
and the sf object (which essentially is just a conventional data.frame with a \code{geometry()} column) containing all the observations (treated and untreated).
The treated indicator variable has to be assigned before (potentially with \code{\link{assign_treated}}) and be part of the sf object as a column.
}
\details{
This function nests \code{\link[rdrobust]{rdrobust}}. All its options (aside from running variable \code{x} and cutoff \code{c}) are available here as well (e.g. bw selection, cluster level, kernel, weights).
Check the documentation in the \code{rdrobust} package for details. (bandwidth selection default in \code{rdrobust} is bwselect = 'mserd')

To visualise the output, use \code{\link{plotspatialrd}} for a graphical representation. You can use \code{\link{printspatialrd}} (or an R package of your choice) for a table output. .
}
\references{
Calonico, Cattaneo and Titiunik (2014): Robust Nonparametric Confidence Intervals for Regression-Discontinuity Designs, Econometrica 82(6): 2295-2326.
}
