% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialrd.R
\name{spatialrd}
\alias{spatialrd}
\title{non-parametric Spatial RD / GRD}
\usage{
spatialrd(
  y,
  data,
  cutoff.points,
  treated,
  minobs = 50,
  bwfix_m = NA,
  sample = F,
  samplesize = NA,
  sparse.exclusion = F,
  spatial.object = T,
  RATestvec = NA,
  ...
)
}
\arguments{
\item{y}{The name of the dependent variable in the points frame in the form of a string}

\item{data}{sf data.frame with points that describe the observations}

\item{cutoff.points}{sf object of borderpoints (provided by user or obtained with \code{\link{discretise_border}})}

\item{treated}{column that contains the treated dummy (as string)}

\item{minobs}{the minimum amount of observations in each estimation for the point estimate to be included (default is 50)}

\item{bwfix_m}{fixed bandwidth in meters}

\item{sample}{draw a random sample of points T/F}

\item{samplesize}{if random, how many points}

\item{spatial.object}{return a spatial object (needed if you want to plot the point estimates on a map)?}

\item{RATestvec}{vector with strings of basline covariates in the data object in case the RATest by Canay, Kamat (2018) is desired}
}
\value{
a data.frame or spatial data.frame (sf object) in case spatial.object = T (default)
}
\description{
This function loops over all boundary points and locally estimates a non-parametric RD (local linear regression as usual)
using the rdrobust function from the rdrobust package from Calonico, Cattaneo, Titiunik (2014) Econometrica.
It takes in the discretised cut-off point file (the RDcut-off linestring chopped into parts by the \code{\link{discretise_border}} function)
and the sf object (which essentially is just a conventional data.frame with a geometry() column) containing all the observations (treated and untreated).
The treated dummy variable has to be assigned before (potentially with \code{\link{assign_treated}}) and be part of the sf object as a column.
}
\details{
This function nests \code{\link[rdrobust]{rdrobust}}. All its options (aside from running variable \code{x} and cutoff \code{c}) are available here as well (e.g. cluster level, kernel, weights).
Check the documentation in the \code{rdrobust} package for details.

To visualise the output table, use \code{\link{printspatialrd}} or \code{\link{plotspatialrd}}.
}
\examples{
results <- spatialrd(y = "education", data = points_samp.sf, cutoff.points = borderpoints.sf, treated = "treated", minobs = 10, spatial.object = F)
}
